authorization implemented

+++ src/api/notes/handler.js
@@ -13,8 +13,9 @@ class NotesHandler {
       async postNoteHandler(request, h) {
         this._validator.validateNotePayload(request.payload);
         const { title = 'untitled', body, tags } = request.payload;
-    
-        const noteId = await this._service.addNote({ title, body, tags });
+        const { id: credentialId } = request.auth.credentials;
+
+        const noteId = await this._service.addNote({ title, body, tags, credentialId});
     
         const response = h.response({
           status: 'success',
@@ -28,7 +29,8 @@ class NotesHandler {
       }
     
       async getNotesHandler() {
-        const notes = await this._service.getNotes();
+        const { id: credentialId } = request.auth.credentials;
+        const notes = await this._service.getNotes(credentialId);
         return {
           status: 'success',
           data: {
@@ -39,6 +41,10 @@ class NotesHandler {
     
       async getNoteByIdHandler(request, h) {
         const { id } = request.params;
+        const { id: credentialId } = request.auth.credentials;
+
+        
+        await this._service.verifyNoteOwner(id, credentialId);
         const note = await this._service.getNoteById(id);
         return {
           status: 'success',
@@ -51,8 +57,10 @@ class NotesHandler {
       putNoteByIdHandler(request, h) {
         this._validator.validateNotePayload(request.payload);
         const { id } = request.params;
-    
-          this._service.editNoteById(id, request.payload);
+        const { id: credentialId} = request.auth.credentials;
+
+        this._service.verifyNoteOwner(id, credentialId);
+        this._service.editNoteById(id, request.payload);
     
         return {
           status: 'success',
@@ -62,6 +70,9 @@ class NotesHandler {
     
       deleteNoteByIdHandler(request, h) {
         const { id } = request.params;
+        const { id: credentialId } = request.auth.credentials;
+
+        this._service.verifyNoteOwner(id, credentialId);
         this._service.deleteNoteById(id);
     
         return {
